Split ThemeSetup into 6 focused classes; decouple Settings from parent
ThemeSetup was a 364-line God class that mixed hook wiring, hook
callbacks, and embedded HTML fragments. Split it into:

- ThemeBootstrap   - theme supports, image sizes, asset enqueueing,
                     admin-bar visibility, Genesis lib/init.php boot
- NavCustomizations- primary menu placement, be_follow_icons filter,
                     footer menu
- PostDisplay      - post info filter, author avatars, author box,
                     featured-image archive placement, gravatar
- HomepageWidgets  - 4 home sidebars + home_large_featured /
                     home_featured_widgets renderers
- SiteChrome       - favicons, footer copyright, login bar
- SharingControls  - hide-sharing-on-protected-pages + WC sidebar removal

Each class registers its own hooks in the constructor. functions.php
now instantiates the six classes directly, no more empty constructor +
separate init() dance.

Migrated tests/test-ThemeSetup.php into one file per class so each
test targets the class that actually owns the behavior.

Separately, Settings no longer holds a back-reference to
PowerOfFamiliesPrograms. It takes the token string and the field
renderer (AdminAPI) directly, so register_settings() no longer has
to traverse $this->parent->admin. PowerOfFamiliesPrograms constructs
AdminAPI first (when in admin context) and injects it into Settings.

My_Programs followed the same pattern - took $parent only to read
$parent::TOKEN - so it now takes the token string directly too.

No behavior changes intended; this is a structural refactor.
