--- wp-content/themes/power-of-families/inc/PowerOfFamilies/Avanti/ThemeSetup.php
@@ -1,364 +1,0 @@
-<?php
-
-namespace PowerOfFamilies\Avanti;
-
-
-class ThemeSetup
-{
-    const RUNNING_PROD = 2 ** 0;
-    const RUNNING_DEV = 2 ** 1;
-    public int $run_location;
-
-    public function __construct()
-    {
-    }
-
-    public function init(): void
-    {
-        $this->child_theme_setup();
-        $this->hideAdminBarFromSubscribers(wp_get_current_user());
-        $this->display_author_box_on_single_posts();
-    }
-
-    public function child_theme_setup(): void
-    {
-        if (!defined('PHPUNIT_RUNNING')) {
-            require_once get_template_directory() . '/lib/init.php';
-        }
-        $this->register_theme_support();
-
-        /* Move primary menu into header */
-        //https://wpbeaches.com/switching-primary-menu-genesis-theme-header-right/
-        remove_action('genesis_after_header', 'genesis_do_nav');
-        add_action('genesis_header_right', 'genesis_do_nav');
-        add_theme_support('genesis-structural-wraps', array('header', 'menu-secondary', 'footer-widgets', 'footer')); //menu-primary is removed
-        unregister_sidebar('header-right');
-
-        /* Filter primary menu */
-        add_filter('genesis_nav_items', [$this, 'be_follow_icons'], 10, 2);
-        add_filter('wp_nav_menu_items', [$this, 'be_follow_icons'], 10, 2);
-        add_filter('wp_nav_menu_header_items', [$this, 'be_follow_icons'], 10, 2);
-
-        //edit the way the post info displays
-        add_filter('genesis_post_info', [$this, 'sp_post_info_filter']);
-
-        // Remove Post Meta
-        remove_action('genesis_entry_footer', 'genesis_post_meta');
-
-        /* Display Post Author Avatars */
-        add_action('genesis_entry_header', [$this, 'wpsites_post_author_avatars']);
-
-        // Move featured image in archives
-        remove_action('genesis_entry_content', 'genesis_do_post_image', 8);
-        add_action('genesis_entry_header', 'genesis_do_post_image', 1);
-
-        // Add new archive image size
-        add_image_size('archive', 170, 170, true);
-        add_image_size('ubermenu', 137, 137, true);
-        add_image_size('sidebar', 300, 300, true);
-        add_image_size('featured-posts', 60, 60, true);
-
-        // Change default avatar
-        add_filter('avatar_defaults', [$this, 'newgravatar']);
-
-        // Add Footer menu
-        add_action('genesis_footer', [$this, 'power_of_families_footer_menu'], 0);
-
-        // Add Copyright to the footer
-        remove_action('genesis_footer', 'genesis_do_footer');
-        add_action('genesis_footer', [$this, 'power_of_families_footer']);
-
-        // Add Favicons
-        add_action('wp_head', [$this, 'pre_load_favicon']);
-        add_action('admin_head', [$this, 'pre_load_favicon']);
-
-        // Setup Widget areas
-        add_action('widgets_init', [$this, 'createWidgets']);
-
-        // Add login bar
-        if (!is_user_logged_in()) {
-            add_action('genesis_before_header', [$this, 'login_bar'], 10);
-        }
-
-        // Hide Sharing buttons on protected pages
-        // @todo: Need to update this for Groups instead of wishlist member
-        add_filter('get_post_metadata', [$this, 'hide_on_protected_pages'], 10, 4);
-        add_filter('get_page_metadata', [$this, 'hide_on_protected_pages'], 10, 4);
-
-        // Remove sidebar on single product pages
-        add_action('wp', function () {
-            if (is_product()) {
-                add_filter('genesis_pre_get_option_site_layout', '__genesis_return_full_width_content');
-            }
-        });
-
-        // Add Javascript and stylesheets
-        add_action('wp_enqueue_scripts', [$this, 'custom_load_styles_and_scripts'], 0);
-    }
-
-    public function custom_load_styles_and_scripts(): void
-    {
-        $js_asset = require get_theme_file_path('dist/main.ts.asset.php');
-        wp_enqueue_script(
-            'pof_theme_scripts',
-            get_stylesheet_directory_uri() . '/dist/main.ts.js',
-            $js_asset['dependencies'],
-            $js_asset['version']
-        );
-
-        wp_enqueue_style('custom-google-fonts', 'https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,300;0,400;0,500;0,600;0,700;1,300;1,400&family=Playfair+Display:wght@400;700&display=swap', false);
-        wp_enqueue_style(
-            'power_of_families_styles',
-            get_stylesheet_directory_uri() . '/dist/main.ts.css',
-            [],
-            $js_asset['version']
-        );
-    }
-
-    public function register_theme_support(): void
-    {
-        $theme_supports = genesis_get_config('theme-supports');
-
-        foreach ($theme_supports as $feature => $args) {
-            add_theme_support($feature, $args);
-        }
-    }
-
-    public function power_of_families_footer(): void
-    {
-?>
-        <p>&copy; Copyright 2017 - <?php echo esc_html( date( 'Y' ) ) ?> <a href="https://poweroffamilies.com">Power of Families</a>
-    <?php
-    }
-
-    public function power_of_families_footer_menu(): void
-    {
-        $args = array(
-            'theme_location' => 'tertiary',
-            'container' => 'nav',
-            'container_class' => 'wrap',
-            'menu_class' => 'menu genesis-nav-menu menu-tertiary',
-            'depth' => 1,
-        );
-        echo '<div class="footer-menu-container">';
-        wp_nav_menu($args);
-        echo '</div>';
-    }
-
-    public function newgravatar(array $avatar_defaults): array
-    {
-        $myavatar = get_stylesheet_directory_uri() . '/assets/images/default_avatar.jpg';
-        $avatar_defaults[$myavatar] = "Power of Families Avatar";
-
-        return $avatar_defaults;
-    }
-
-    public function wpsites_post_author_avatars(): void
-    {
-        if (is_single()) {
-            echo get_avatar(get_the_author_meta('email'), 60);
-        }
-    }
-
-    /**
-     * Hide the admin bar for subscribers.
-     *
-     * @param \WP_User|null $current_user The current user object.
-     */
-    public function hideAdminBarFromSubscribers(?\WP_User $current_user = null): bool
-    {
-        if ($current_user && $current_user->exists()) {
-            $user_roles = $current_user->roles;
-            if (in_array('subscriber', $user_roles)) {
-                show_admin_bar(false);
-                return false;
-            }
-        }
-        return true;
-    }
-
-    public function be_follow_icons($menu, $args): string
-    {
-        //Top menu
-        $args = (array)$args;
-        if ('primary' !== $args['theme_location']) {
-            return $menu;
-        }
-        if (is_user_logged_in()) {
-            $follow = '<li class="menu-item menu-item-has-children" ><a href="/my-account/" title="My Account">My Account</a>';
-            //$follow .= '<div class="sub-menu-toggle-container"><button class="sub-menu-toggle" role="button" aria-pressed="false"><span class="hide-activated">Open Navigation</span><span class="hide-deactivated">Close Navigation</span></button></div>';
-            $follow .= '<ul class="sub-menu">';
-            $follow .= '<li class="menu-item menu-login-link"><a href="/my-programs/" title="Go to My Programs">My Programs</a></li>';
-            $follow .= '<li class="menu-item"><a href="/wp-admin/profile.php" title="Update My Profile">Update My Profile</a></li>';
-            $follow .= '<li class="menu-item"><a href="' . wp_logout_url(get_permalink()) . '" title="Logout">Logout</a></li>';
-            $follow .= '</ul></li>';
-        } else {
-            $follow = '<li class="menu-item menu-login-link" data-toggle="collapse" data-target="#login-bar"><a href="#">Log In</a></li>';
-        }
-
-        return $menu . $follow;
-    }
-
-
-    public function pre_load_favicon(): void
-    {
-        $favicon_directory = get_stylesheet_directory_uri() . '/assets/images/favicon/';
-
-        echo '<link rel="shortcut icon" href="' . $favicon_directory . 'favicon.ico">' . PHP_EOL
-            . '<link rel="apple-touch-icon-precomposed" sizes="57x57"   href="' . $favicon_directory . 'apple-touch-icon-57x57.png" />' . PHP_EOL
-            . '<link rel="apple-touch-icon-precomposed" sizes="114x114" href="' . $favicon_directory . 'apple-touch-icon-114x114.png" />' . PHP_EOL
-            . '<link rel="apple-touch-icon-precomposed" sizes="72x72"   href="' . $favicon_directory . 'apple-touch-icon-72x72.png" />' . PHP_EOL
-            . '<link rel="apple-touch-icon-precomposed" sizes="144x144" href="' . $favicon_directory . 'apple-touch-icon-144x144.png" />' . PHP_EOL
-            . '<link rel="apple-touch-icon-precomposed" sizes="60x60"   href="' . $favicon_directory . 'apple-touch-icon-60x60.png" />' . PHP_EOL
-            . '<link rel="apple-touch-icon-precomposed" sizes="120x120" href="' . $favicon_directory . 'apple-touch-icon-120x120.png" />' . PHP_EOL
-            . '<link rel="apple-touch-icon-precomposed" sizes="76x76"   href="' . $favicon_directory . 'apple-touch-icon-76x76.png" />' . PHP_EOL
-            . '<link rel="apple-touch-icon-precomposed" sizes="152x152" href="' . $favicon_directory . 'apple-touch-icon-152x152.png" />' . PHP_EOL
-            . '<link rel="icon" type="image/png" href="' . $favicon_directory . 'favicon-196x196.png" sizes="196x196" />' . PHP_EOL
-            . '<link rel="icon" type="image/png" href="' . $favicon_directory . 'favicon-96x96.png" sizes="96x96" />' . PHP_EOL
-            . '<link rel="icon" type="image/png" href="' . $favicon_directory . 'favicon-32x32.png" sizes="32x32" />' . PHP_EOL
-            . '<link rel="icon" type="image/png" href="' . $favicon_directory . 'favicon-16x16.png" sizes="16x16" />' . PHP_EOL
-            . '<link rel="icon" type="image/png" href="' . $favicon_directory . 'favicon-128.png" sizes="128x128" />';
-    }
-
-    public function display_author_box_on_single_posts(): void
-    {
-        add_filter('get_the_author_genesis_author_box_single', '__return_true');
-        remove_action('genesis_after_entry', 'genesis_do_author_box_single', 8);
-        add_action('genesis_entry_content', 'genesis_do_author_box_single', 10);
-    }
-
-
-    public function sp_post_info_filter(string $post_info): string
-    {
-        if (is_single()) {
-            $post_info = 'by [post_author_posts_link] on [post_date format="M j, Y"] [post_comments] [post_edit] [post_categories sep=", " before="Posted in: "]';
-        } else {
-            $post_info = 'by [post_author_posts_link] on [post_date format="M j, Y"]';
-        }
-        return $post_info;
-    }
-
-    public function createWidgets(): void
-    {
-        genesis_register_sidebar(array(
-            'id' => 'home_large_featured',
-            'name' => __('Home Large Feature', 'powerofmoms'),
-            'description' => __('This is the large home section.', 'news'),
-        ));
-        genesis_register_sidebar(array(
-            'id' => 'home_left',
-            'name' => __('Home Left Feature', 'powerofmoms'),
-            'description' => __('This is the left home featured content section.', 'news'),
-        ));
-        genesis_register_sidebar(array(
-            'id' => 'home_middle',
-            'name' => __('Home Middle Feature', 'powerofmoms'),
-            'description' => __('This is the middle home featured content section.', 'news'),
-        ));
-        genesis_register_sidebar(array(
-            'id' => 'home_right',
-            'name' => __('Home Right Feature', 'powerofmoms'),
-            'description' => __('This is the right home featured content section.', 'news'),
-        ));
-
-        add_action('genesis_before_content', [$this, 'home_large_featured']);
-        add_action('genesis_before_content', [$this, 'home_featured_widgets']);
-    }
-
-    public function home_large_featured(): void
-    {
-        if (is_home()) {
-            genesis_widget_area('home_large_featured', array(
-                'before' => '<div class="home-large-featured"><div class="wrap">',
-                'after' => '</div></div>',
-            ));
-        }
-    }
-
-
-    public function home_featured_widgets(): void
-    {
-        if (is_home()) {
-            echo '<div class="home-featured-widgets"><div class="wrap"><h2 class="panel-title">What\'s New</h2>';
-            genesis_widget_area('home_left', array(
-                'before' => '<div class="home-featured-widget-1 widget-area home-featured-widget"><div class="home-featured-widget-inner">',
-                'after' => '</div></div>',
-            ));
-            genesis_widget_area('home_middle', array(
-                'before' => '<div class="home-featured-widget-2 widget-area home-featured-widget"><div class="home-featured-widget-inner">',
-                'after' => '</div></div>',
-            ));
-            genesis_widget_area('home_right', array(
-                'before' => '<div class="home-featured-widget-3 widget-area home-featured-widget"><div class="home-featured-widget-inner">',
-                'after' => '</div></div>',
-            ));
-            echo '</div></div>';
-        }
-    }
-
-    public function login_bar(): void
-    {
-
-        echo '<div class="login-bar collapse" id="login-bar"><div class="wrap">';
-        $request = isset( $_GET['wlfrom'] )
-            ? esc_url_raw( wp_validate_redirect( sanitize_url( wp_unslash( $_GET['wlfrom'] ) ), home_url() ) )
-            : esc_url_raw( wp_unslash( $_SERVER['REQUEST_URI'] ) );
-        $args = array(
-            'echo' => true,
-            //'redirect'       => $request,
-            'redirect' => '\/my-programs\/',
-            'form_id' => 'loginform',
-            'label_username' => __('Username'),
-            'label_password' => __('Password'),
-            'label_remember' => __('Remember Me'),
-            'label_log_in' => __('Log In'),
-            'id_username' => 'user_login',
-            'id_password' => 'user_pass',
-            'id_remember' => 'rememberme',
-            'id_submit' => 'wp-submit',
-            'remember' => true,
-            'value_username' => null,
-            'value_remember' => true
-        );
-        echo '<h2>Member Login</h2>';
-        echo wp_login_form($args);
-        echo '<div class="register"><a href="/register" class="button">Register</a></div>';
-        printf(
-            '<div class="lost_password"><a href="%s" title="Lost Password">Lost Password?</a></div>',
-            wp_lostpassword_url(get_permalink())
-        );
-
-        echo '<button type="button" class="close" data-toggle="collapse" data-target="#login-bar">
-<span aria-hidden="true">&times;</span><span class="sr-only">Close</span></button></div></div>';
-    }
-
-    private function get_protected_pages(): array
-    {
-        if (false === ($ids = get_transient('pof_protected_pages'))) {
-            $ids = [];
-            if (function_exists('wlmapi_get_protected_pages')) {
-                $pages = wlmapi_get_protected_pages();
-                $ids = array_merge($ids, array_map(function ($item) {
-                    return $item->ID;
-                }, $pages['pages']['page']));
-            }
-            if (function_exists('wlmapi_get_protected_posts')) {
-                $pages = wlmapi_get_protected_posts();
-                $ids = array_merge($ids, array_map(function ($item) {
-                    return $item->ID;
-                }, $pages['posts']['post']));
-            }
-            $ids = array_map('intval', $ids);
-            set_transient('pof_protected_pages', $ids, HOUR_IN_SECONDS);
-        }
-        return $ids;
-    }
-
-    public function hide_on_protected_pages($metadata, $object_id, $meta_key, $single): mixed
-    {
-        if ($meta_key === 'essb_off' && in_array($object_id, $this->get_protected_pages())) {
-            return 'true';
-        }
-        return $metadata;
-    }
-}
